+++ aeron-client/src/main/c/collections/aeron_array_to_ptr_hash_map.h
@@ -58,7 +58,7 @@ inline uint64_t aeron_array_hash(const uint8_t *arr, size_t length)
 
 inline size_t aeron_array_to_ptr_hash_map_hash_key(uint64_t key, size_t mask)
 {
-    return (key * 31) & mask;
+    return (size_t)((key * 31) & mask);
 }
 
 inline bool aeron_array_to_ptr_hash_map_compare(
