+++ aeron-client/src/main/c/collections/aeron_str_to_ptr_hash_map.h
@@ -50,7 +50,7 @@ aeron_str_to_ptr_hash_map_t;
 
 inline size_t aeron_str_to_ptr_hash_map_hash_key(uint64_t key, size_t mask)
 {
-    return (key * 31) & mask;
+    return (size_t)((key * 31) & mask);
 }
 
 inline bool aeron_str_to_ptr_hash_map_compare(
